Fix client.ts file
`client.ts` is autogenerated.

The file included manual changes (#607) that were overwritten by the
master build.

The PR build would not overwrite change, hence previous PR builds were
passing.

This change reverts manual changes and ensures the PR build also
regenerates `client.ts`.
