+++ package.json
@@ -43,7 +43,7 @@
   },
   "devDependencies": {
     "@lune-climate/eslint-config": "git+https://github.com/lune-climate/eslint-config.git#master",
-    "@lune-climate/openapi-typescript-codegen": "^0.1.11",
+    "@lune-climate/openapi-typescript-codegen": "^0.1.12",
     "typescript": "^5.1.6",
     "@types/node": "^20.4.5",
     "@typescript-eslint/eslint-plugin": "^7.0.1",
