+++ src/client.ts
@@ -1,8 +1,12 @@
-import axios, { AxiosError, AxiosInstance, AxiosResponse, isAxiosError } from 'axios'
+import axios, { AxiosInstance, AxiosResponse, isAxiosError } from 'axios'
 import camelCaseKeys from 'camelcase-keys'
 
 import { ClientConfig } from './core/ClientConfig.js'
-import { Meta, Method, Methods } from './core/SuccessResponse.js'
+import {
+    ExtendedAxiosError,
+    ExtendedAxiosResponse,
+    extractRequestFromResponseInterceptor,
+} from './core/SuccessResponse.js'
 import { AccountsService } from './services/AccountsService.js'
 import { AnalyticsService } from './services/AnalyticsService.js'
 import { BundlePortfoliosService } from './services/BundlePortfoliosService.js'
@@ -28,39 +32,6 @@ function applyMixins(derivedCtor: any, constructors: any[]) {
             )
         })
     })
-}
-
-export interface ExtendedAxiosResponse<T = any> extends AxiosResponse<T> {
-    _meta: Meta<T>
-}
-
-interface ExtendedAxiosError<T = any> extends AxiosError<T> {
-    response?: ExtendedAxiosResponse<T>
-}
-
-function extractRequestFromResponseInterceptor(response: AxiosResponse): {
-    request: object | null
-    method: Method
-    url: string
-    requestHeaders: { contentType: string | null }
-} {
-    const req = response.config
-    const data = req.data
-
-    if (!Methods.includes((req.method ?? '').toLowerCase())) {
-        throw new Error(`Unexpected method: ${req.method}`)
-    }
-
-    return {
-        method: req.method!,
-        url: req.baseURL!,
-        request: !data ? null : typeof data === 'string' ? JSON.parse(data) : data,
-        requestHeaders: {
-            contentType: req.headers['Content-Type']
-                ? (req.headers['Content-Type'] as string)
-                : null,
-        },
-    }
 }
 
 export class LuneClient {
